refactor(tests): move array expression tests to `test_value_exprs.py`
